+++ test/AuctionHouse/sendPayment.t.sol
@@ -20,7 +20,7 @@ contract SendPaymentTest is Test, Permit2User {
     address internal OWNER = address(0x3);
 
     // Function parameters
-    uint256 internal lotId = 1;
+    uint96 internal lotId = 1;
     uint256 internal paymentAmount = 1e18;
     MockFeeOnTransferERC20 internal quoteToken;
     MockHook internal hook;
